Update initial admin permissions periodically
This updates the maintenance service to update the permissions of
initial admins. This ensures that the initial admin always has "all"
permissions, even across software upgrades. It adds telemetry to the
maintenance service, and also adjusts the service to run once as soon
as the server is started (in addition to every midnight).

Fix: https://github.com/io7m/idstore/issues/27